Move public class to its own file
