+++ collect_app/src/main/java/org/odk/collect/android/database/DatabaseConnection.kt
@@ -1,13 +1,10 @@
 package org.odk.collect.android.database
 
 import android.content.Context
-import android.content.ContextWrapper
-import android.database.DatabaseErrorHandler
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteDatabase.CursorFactory
 import android.database.sqlite.SQLiteOpenHelper
 import timber.log.Timber
-import java.io.File
 
 /**
  * Allows access to a database file. The actual underlying connecting to this database will be
@@ -95,30 +92,3 @@ private class DatabaseMigratorSQLiteOpenHelper(
     }
 }
 
-/**
- * Allows creating an [SQLiteOpenHelper] that references a `.db` somewhere other than the standard
- * Android database path (by passing this as [Context] to the constructor).
- */
-class AltDatabasePathContext(private val path: String, context: Context) : ContextWrapper(context) {
-
-    override fun getDatabasePath(name: String): File {
-        return File(path + File.separator + name)
-    }
-
-    override fun openOrCreateDatabase(
-        name: String,
-        mode: Int,
-        factory: CursorFactory,
-        errorHandler: DatabaseErrorHandler?
-    ): SQLiteDatabase {
-        return openOrCreateDatabase(name, mode, factory)
-    }
-
-    override fun openOrCreateDatabase(
-        name: String,
-        mode: Int,
-        factory: CursorFactory
-    ): SQLiteDatabase {
-        return SQLiteDatabase.openOrCreateDatabase(getDatabasePath(name), null)
-    }
-}
